moved to Flink 2 and it's Guava dependency
