+++ pom.xml
@@ -60,9 +60,9 @@
 		<maven.compiler.target>${java.version}</maven.compiler.target>
 		<target.java.version>${java.version}</target.java.version>
 
-		<flink.version>1.20.0</flink.version>
-		<kryo.version>2.24.0</kryo.version>
-		<objenesis.version>2.1</objenesis.version>
+		<flink.version>2.0.0</flink.version>
+		<kryo.version>5.6.2</kryo.version>
+		<objenesis.version>3.3</objenesis.version>
 
 		<snowflake-ingest.version>2.3.0</snowflake-ingest.version>
 		<assertj.version>3.21.0</assertj.version>
@@ -70,6 +70,7 @@
 		<testcontainers.version>1.19.1</testcontainers.version>
 		<mockito.version>5.2.0</mockito.version>
 		<jackson-bom.version>2.15.3</jackson-bom.version>
+		<guava.version>32.1.3-jre-19.0</guava.version>
 
 		<japicmp.skip>false</japicmp.skip>
 		<japicmp.referenceVersion>3.0.0-1.16</japicmp.referenceVersion>
@@ -108,12 +109,13 @@
 			<scope>provided</scope>
 		</dependency>
 
-		<!-- Test dependencies -->
 		<dependency>
-			<groupId>org.junit.jupiter</groupId>
-			<artifactId>junit-jupiter</artifactId>
-			<scope>test</scope>
+			<groupId>org.apache.flink</groupId>
+			<artifactId>flink-shaded-guava</artifactId>
+			<version>${guava.version}</version>
 		</dependency>
+
+		<!-- Test dependencies -->
 
 		<dependency>
 			<groupId>org.assertj</groupId>
@@ -267,7 +269,7 @@
 
 			<!-- For dependency convergence -->
 			<dependency>
-				<groupId>com.esotericsoftware.kryo</groupId>
+				<groupId>com.esotericsoftware</groupId>
 				<artifactId>kryo</artifactId>
 				<version>${kryo.version}</version>
 			</dependency>
